// modules/transaction-manager/modules/signature-tracking/enum/signature-tracking.enum.ts
export enum SignatureTrackingTableKeys {
  DATE = 'creationDate',
  DATA_ORDER = 'interfaceDate',
  TRANSACTION = 'serviceDescription',
  REFERENCE = 'reference',
  CURRENCY = 'currency',
  AMOUNT = 'amount',
  USER = 'creator',
  ERROR_DETAIL = 'errorDetail',
  FILTER_SEARCH = 'search',
  ACCOUNT = 'account',
  USER_CREATED = 'user',
  USER_NAME = 'userName',
  DATE_TIME = 'dateTime'
}


export const subTitleMessage = {
  delete: 'delete_transaction',
  return: 'rejection_transaction',
  process: 'process_transaction',
  authorize: 'authorizes_transaction',
  send: 'send_transaction',
};